Fix sourcing of asdf.sh
Installing asdf via Homebrew now says:

```
To use asdf, add the following line to your ~/.zshrc:
  . /opt/homebrew/opt/asdf/libexec/asdf.sh
```

Yes, the laptop script is doing:

`. /opt/homebrew/opt/asdf/asdf.sh`

This used to work, but in no longer does because of the change in
https://github.com/Homebrew/homebrew-core/pull/81664

This corrects the file location being source.

This failure was showing in CI and UTM locally for me.